Changed unsigned short to size_t

+++ src/client/systems/background/System+MoveBackground.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include "EwECS/World.hpp"
 #include "SFMLDisplayClass.hpp"
 #include "System.hpp"
@@ -10,7 +11,7 @@ namespace ECS {
         auto deltaTime = Core::World::getInstance().getDeltaTime();
         const auto size = aType.size();
 
-        for (unsigned short i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             if (!aType[i].has_value() || !aType[i].value().isBackground || !aPos[i].has_value()
                 || !aSpeed[i].has_value()) {
                 continue;
